requirements.txt
asgiref==3.8.1
diff-match-patch==20241021
Django==5.1.4
django-allauth==65.3.1
django-ckeditor==6.7.2
django-ckeditor-5==0.2.15
django-cors-headers==4.6.0
django-crispy-forms==2.3
django-debug-toolbar==4.4.6
django-import-export==4.3.3
django-jazzmin==3.0.1
django-js-asset==3.0.1
django-mathfilters==1.0.0
django-taggit==6.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.7.0
mysqlclient==2.2.6
pillow==11.0.0
psycopg2==2.9.10
PyJWT==2.10.1
pymongo==4.10.1
shortuuid==1.0.13
sqlparse==0.5.2
tablib==3.7.0
tzdata==2024.2
uuid==1.30
